Simple Base and Velocity addons for emitter

+++ SobrassadaEngine/Utils/Particles/ParticleAddon.cpp
@@ -1,6 +1,7 @@
 #include "ParticleAddon.h"
 
-ParticleAddon::ParticleAddon(const rapidjson::Value& initialState, ResourceEmitter* owner) : emitterOwner(owner)
+ParticleAddon::ParticleAddon(const rapidjson::Value& initialState, ResourceEmitter* owner)
+    : emitterOwner(owner)
 {
     if (initialState.HasMember("Enabled")) isEnabled = initialState["Enabled"].GetBool();
     if (initialState.HasMember("AddonType")) addonType = ParticleAddonType(initialState["AddonType"].GetInt());
@@ -9,5 +10,5 @@ ParticleAddon::ParticleAddon(const rapidjson::Value& initialState, ResourceEmitt
 void ParticleAddon::Save(rapidjson::Value& targetState, rapidjson::Document::AllocatorType& allocator) const
 {
     targetState.AddMember("Enabled", isEnabled, allocator);
-    targetState.AddMember("AddonType", addonType, allocator);
+    targetState.AddMember("AddonType", (int)addonType, allocator);
 }
